dummyLifeEvents definition moved
